% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_as_table.R
\name{create_as_table}
\alias{create_as_table}
\title{Creates Tables for Particular Years and Countries}
\usage{
create_as_table(alum, cn, year)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function loads Alumnae Survey data and creates formatted tables to be shared with country teams.
}
