% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_as.R
\name{fn_as}
\alias{fn_as}
\title{Estimate Parameters}
\usage{
fn_as(asur, coun, var, year)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Functions in this file generate numbers from cleaned Alumnae Survey data, to be used in graphs and tables.
}
